patcher: do not return load base from ELF load
Better to discover this in load_patch.
Also first_vma + vma_start helpers can be used instead of additional
first_mmap_addr helper.

+++ patcher/include/elf.h
@@ -13,8 +13,8 @@ int parse_elf_binpatch(struct patch_info_s *binpatch, const char *patchfile);
 
 struct process_ctx_s;
 struct elf_info_s;
-int64_t load_elf(struct process_ctx_s *ctx, struct list_head *segments,
-		 const struct elf_info_s *ei, uint64_t hint);
+int load_elf(struct process_ctx_s *ctx, struct list_head *segments,
+	     const struct elf_info_s *ei, uint64_t hint);
 int unload_elf(struct process_ctx_s *ctx, struct list_head *segments);
 
 int elf_create_info(const char *path, struct elf_info_s **elf_info);
